requirements.txt
fastapi
uvicorn[standard]
streamlit
pandas
requests
redis
pyyaml